gateway/requirements.txt
fastapi
uvicorn
aio-pika
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-fastapi
opentelemetry-exporter-jaeger
pydantic